// src/types/pet/pet-enums.ts
export enum PetSpecies {
  DOG = 'dog',
  CAT = 'cat',
  BIRD = 'bird',
  RABBIT = 'rabbit',
  HAMSTER = 'hamster',
  FISH = 'fish',
  REPTILE = 'reptile',
  HORSE = 'horse',
  OTHER = 'other',
}

export enum PetGender {
  MALE = 'male',
  FEMALE = 'female',
  UNKNOWN = 'unknown',
}

export enum PetSize {
  TINY = 'tiny', // < 5 lbs
  SMALL = 'small', // 5-20 lbs
  MEDIUM = 'medium', // 20-50 lbs
  LARGE = 'large', // 50-100 lbs
  GIANT = 'giant', // > 100 lbs
}